Added IF statement to test cd command

+++ backups_home.sh
@@ -2,9 +2,9 @@
 
 tstamp=`date +%Y%m%d`
 cd $HOME/.local/share/gnome-shell/extensions
-tar -zcf $HOME/backups/extensions_${tstamp}.tar.gz ./.
-
-cd $HOME/GIT_REPO
-rm -rf `find . -maxdepth 1 -type d  -mtime +10 -ls`
-
+if [ $? == 0 ]; then
+	tar -zcf $HOME/backups/extensions_${tstamp}.tar.gz ./.
+	cd $HOME/GIT_REPO
+	rm -rf `find . -maxdepth 1 -type d  -mtime +10 -ls`
+fi
 exit 0
